Fix some more narrowing failures.
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3754222
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1023571}

+++ chrome/browser/ui/views/extensions/extensions_tabbed_menu_view_unittest.cc
@@ -38,9 +38,9 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace {
 
 // Combobox option's indexes for site access menu items.
-constexpr int kOnClickComboboxIndex = 0;
-constexpr int kOnSiteComboboxIndex = 1;
-constexpr int kOnAllSitesComboboxIndex = 2;
+constexpr size_t kOnClickComboboxIndex = 0;
+constexpr size_t kOnSiteComboboxIndex = 1;
+constexpr size_t kOnAllSitesComboboxIndex = 2;
 // Button's indexes for site access settings
 constexpr int kGrantAllExtensionsIndex = 0;
 constexpr int kBlockAllExtensionsIndex = 1;
